Split the decorator code into separate files
