+++ decorator.rb
@@ -10,19 +10,3 @@ class Decorator < Nameable
     @namable.correct_name
   end
 end
-
-class CapitalizeDecorator < Decorator
-  def correct_name
-    super.capitalize
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    if super.length > 10
-      super[0..9]
-    else
-      super
-    end
-  end
-end
