+++ person.rb
@@ -1,5 +1,6 @@
 require './nameable'
-require './decorator'
+require './capitalize_decorator'
+require './trimmer_decorator'
 
 class Person < Nameable
   def initialize(age, name = 'Unknown', parent_permission: true)
